Add files via upload
The Color Game:

Once you run the code, it will open on a different window (it might appear as a file in your Taskbar). you can expand the window and once you are ready you can begin the game by pressing Enter. To play the game, write down the COLOR of the word that shows up on your screen instead of the word itself, this game is meant to confuse you! ( you are timed)